+++ packages/platform-ui/src/components/showcase/ContainersScreenShowcase.tsx
@@ -1,4 +1,4 @@
-import ContainersScreen, { Container, ContainerStatus } from '../screens/ContainersScreen';
+import ContainersScreen, { Container, ContainerStatus } from './screens/ContainersScreen';
 
 // Generate sample containers
 const generateSampleContainers = (): Container[] => {
