+++ packages/platform-ui/src/components/showcase/RemindersScreenShowcase.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import RemindersScreen, { Reminder, ReminderStatus } from '../screens/RemindersScreen';
+import RemindersScreen, { Reminder, ReminderStatus } from './screens/RemindersScreen';
 
 // Generate sample reminders
 const generateSampleReminders = (): Reminder[] => {
