+++ packages/platform-ui/src/components/showcase/RunScreenShowcase.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import RunScreen from '../screens/RunScreen';
+import RunScreen from './screens/RunScreen';
 import { RunEvent } from '../RunEventsList';
 import { Status } from '../StatusIndicator';
