+++ packages/platform-ui/src/components/showcase/SecretsScreenShowcase.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import SecretsScreen, { Secret } from '../screens/SecretsScreen';
+import SecretsScreen, { Secret } from './screens/SecretsScreen';
 
 // Generate sample secrets
 const generateSampleSecrets = (): Secret[] => {
