+++ packages/platform-ui/src/components/showcase/VariablesScreenShowcase.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import VariablesScreen, { Variable } from '../screens/VariablesScreen';
+import VariablesScreen, { Variable } from './screens/VariablesScreen';
 
 // Generate sample variables
 const generateSampleVariables = (): Variable[] => {
